Importing Python modules and packages in python programming
b.	You are working on a Python project that requires you to perform various mathematical operations and geometric area calculations. To organize your code better, you decide to create a package named mypackage which includes sub packages pack1 and pack 2 with two modules: mathfunctions and areafunctions Demonstrate the use of the functions by performing a few calculations and printing the results.