Implement python generator and decorators
Write a Python program that includes a generator function to produce a
sequence of numbers. The generator should be able to:
a. Produce a sequence of numbers when provided with start, end, and step
values.
b. Produce a default sequence of numbers starting from 0, ending at 10, and with a step of 1 if no values are provided.